- Use Projective Point serialization for Publickey and Signature. - implement ToBytes and FromBytes from arkworks for Publickey and     Signature. - rename bls_pop.rs to more relevant schnorr_pop.rs
