+++ Cargo.toml
@@ -20,6 +20,8 @@ blake2 = "0.9.2"
 digest = "0.9.0"
 fdh = "0.7.1"
 
+ark-std = { version = "^0.3.0", default-features = false }
+
 ark-ff = { version = "0.3", default-features = false }
 ark-ec = { version = "0.3", default-features = true }
 ark-serialize = { version = "0.3", default-features = false, features = [ "derive" ] }
